Revert "Modify main game loop for use with basic AI unit tests (#3669)"
This reverts commit 1e116be3b078598426fd7f7ff4afe0bd40d78777.
