+++ forge-ai/src/main/java/forge/ai/AiController.java
@@ -519,8 +519,6 @@ public class AiController {
             landList = unreflectedLands;
         }
 
-        // TODO If there's nothing to do with the mana, then play a tapland
-
         //try to skip lands that enter the battlefield tapped
         if (!nonLandsInHand.isEmpty()) {
             CardCollection nonTappedLands = new CardCollection();
