+++ forge-ai/src/main/java/forge/ai/simulation/SpellAbilityPicker.java
@@ -98,13 +98,6 @@ public class SpellAbilityPicker {
             return null;
         }
 
-        if (printOutput && game.getPhaseHandler().getPhase() == PhaseType.MAIN1) {
-            System.out.println("Current state for " + player + ":");
-            for (Card c : player.getCardsIn(ZoneType.Battlefield)) {
-                System.out.println(c);
-            }
-        }
-
         Score origGameScore = new GameStateEvaluator().getScoreForGameState(game, player);
         List<SpellAbility> candidateSAs = getCandidateSpellsAndAbilities();
         if (controller != null) {
@@ -194,7 +187,7 @@ public class SpellAbilityPicker {
 
         SpellAbility bestSa = null;
         Score bestSaValue = origGameScore;
-        print("Evaluating as " + player + "... (orig score = " + origGameScore + ")");
+        print("Evaluating... (orig score = " + origGameScore +  ")");
         for (int i = 0; i < candidateSAs.size(); i++) {
             Score value = evaluateSa(controller, phase, candidateSAs, i);
             if (value.value > bestSaValue.value) {
