+++ forge-core/src/main/java/forge/StaticData.java
@@ -382,10 +382,6 @@ public class StaticData {
         return printSheets;
     }
 
-    /**
-     * Get a database of all non-variant cards
-     * @return
-     */
     public CardDb getCommonCards() {
         return commonCards;
     }
