+++ forge-gui-desktop/src/main/java/forge/view/SimulateMatch.java
@@ -172,7 +172,7 @@ public class SimulateMatch {
         System.out.println("\tq - Quiet flag. Output just the game result, not the entire game log.");
     }
 
-    public static Game simulateSingleGameOfMatch(final Match mc, int timeout) {
+    public static void simulateSingleMatch(final Match mc, int iGame, boolean outputGamelog) {
         final StopWatch sw = new StopWatch();
         sw.start();
 
@@ -182,7 +182,7 @@ public class SimulateMatch {
             TimeLimitedCodeBlock.runWithTimeout(() -> {
                 mc.startGame(g1);
                 sw.stop();
-            }, timeout, TimeUnit.SECONDS);
+            }, 120, TimeUnit.SECONDS);
         } catch (TimeoutException e) {
             System.out.println("Stopping slow match as draw");
         } catch (Exception | StackOverflowError e) {
@@ -195,14 +195,6 @@ public class SimulateMatch {
                 g1.setGameOver(GameEndReason.Draw);
             }
         }
-
-        return g1;
-    }
-
-    public static void simulateSingleMatch(final Match mc, int iGame, boolean outputGamelog) {
-        final StopWatch sw = new StopWatch();
-        sw.start();
-        final Game g1 = simulateSingleGameOfMatch(mc, 120);
 
         List<GameLogEntry> log;
         if (outputGamelog) {
